+++ src/store/user/userInfo.ts
@@ -1,10 +1,10 @@
-
-import { defineStore, type Pinia, storeToRefs } from "pinia";
+import { defineStore, storeToRefs } from "pinia";
 import { computed, ref } from "vue";
 
 import { Currencies } from "../../models/enums/currencies";
 import { BettingPlayerSettings } from "../../models/player";
 import type { IUserData, IUserStatus } from "../../models/user";
+import { ISeasonStartPoints } from "../../models/user";
 import type { IPlayerStats, ISubscriptions, IUserSettings } from "../../services/api/DTO/playerDTO";
 import { useCommon } from "../common";
 
@@ -66,6 +66,7 @@ export const useUserInfo = defineStore("userInfo", () => {
         oddsTypes: [],
         selectedOddsType: "european",
     });
+    const userStartSeasonInfo = ref<ISeasonStartPoints>();
 
     const getUserInfo = computed(() => info.value);
     const getUserBettingBonuses = computed(() => bettingBonuses.value);
@@ -87,6 +88,7 @@ export const useUserInfo = defineStore("userInfo", () => {
     const getUserNickName = computed(() => info.value.nickname || info.value.email);
     const getIsLoadedUsedData = computed(() => isLoadedUsedData.value);
     const getPlayerStats = computed(() => stats.value);
+    const getUserStartSeasonInfo = computed(() => userStartSeasonInfo.value);
 
     function setPlayerStats(data: IPlayerStats) {
         stats.value = data;
@@ -187,6 +189,10 @@ export const useUserInfo = defineStore("userInfo", () => {
         settings.value = data;
     }
 
+    function setUserSeasonStartPoints(data: ISeasonStartPoints) {
+        userStartSeasonInfo.value = data;
+    }
+
     return {
         getUserInfo,
         setUserInfo,
@@ -212,6 +218,7 @@ export const useUserInfo = defineStore("userInfo", () => {
         getIsLoadedUsedData,
         getUserBettingBonuses,
         getPlayerStats,
+        getUserStartSeasonInfo,
         setPlayerStats,
         getSubunitsToUnitsByCode,
         toggleUserIsLogged,
@@ -226,5 +233,6 @@ export const useUserInfo = defineStore("userInfo", () => {
         bettingPlayerSettings,
         setBettingPlayerSettings,
         setUserSettings,
+        setUserSeasonStartPoints,
     };
 });